Moved script tag down, added src
When I try to run the solution in for this lab in qa-internal, I keep getting the error "Cannot read property remove of null". The reason is because the script is loaded before the HTML runs, so querySelector is unable to find "main#main", since it doesn't exist yet. First, I added "src="index.js" to the script tag, then I moved the script tag down to the end of the body. This way, the script will run properly and students will be able to remove the "main#main" element.